Add files via upload
This is the first notebook for the detector, a version 0.1. This is a basic detector built on top of the code from https://github.com/sahupankaj10/Chest-X-ray-image-classification
The CNN is built around the concept of skip connections to reduce the number of parameters. It consist of 5 blocks (can be extended when required). Each block has 3 arms; arm 1 has three convolution of the input (3 by 3 followed by 5 by 5 followed by 7 by 7)
arm 2 has two convolutions on input, 5 by5 and 7 by 7  and arm 3 has just one 7 by 7 convolution. Arm 1 and 2 are concatenated to create arm12; arm12 and 3 are concatenated to create the final conv output. The output is passed through max pool and batch norm layers.
We have two fully connected layers (number can be increased or decreased)